requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Geospatial Data (OPTIONAL for multi-spectral .tif files)
rasterio>=1.5.0
# gdal>=3.7.0  # Note: Requires system-level installation

# Deep Learning Utilities
timm>=0.9.0  # Vision Transformer models
einops>=0.6.1  # Tensor operations

# Training & Optimization
pytorch-msssim>=1.0.0  # SSIM loss
tensorboard>=2.13.0  # Training visualization
tqdm>=4.65.0  # Progress bars

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0  # Advanced config management

# Deployment
onnx>=1.14.0
onnxruntime-gpu>=1.15.0  # Use onnxruntime for CPU-only
onnxslim>=0.1.84

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0  # Interactive plots

# Utilities
python-dotenv>=1.0.0
rich>=13.5.0  # Pretty terminal output
loguru>=0.7.0  # Better logging

# Development (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
